Add mjød and cider to stock update script

+++ api/beers/management/commands/update_stock_from_vmp.py
@@ -4,9 +4,11 @@ from beers.models import Beer, ExternalAPI, Store, Stock
 from django.core.management.base import BaseCommand
 
 
-def call_api(url, store_id, page):
+def call_api(url, store_id, page, product):
     query = (
-        ":name-asc:visibleInSearch:true:mainCategory:øl:availableInStores:"
+        ":name-asc:visibleInSearch:true:mainCategory:"
+        + product
+        + ":availableInStores:"
         + str(store_id)
         + ":"
     )
@@ -38,51 +40,61 @@ class Command(BaseCommand):
         stores_updated = 0
         n = options["stores"]
         stores = Store.objects.all().order_by("store_stock_updated")[:n]
+        products = ["øl", "sider", "mjød"]
 
         for store in stores.iterator():
             stocked_beer = []
 
-            response, total_pages = call_api(url, store.store_id, 0)
-
-            # Update all beers in stock
-            for page in range(0, int(total_pages)):
+            for product in products:
                 try:
-                    response, total_pages = call_api(url, store.store_id, page)
+                    response, total_pages = call_api(url, store.store_id, 0, product)
+                except:
+                    break
 
-                    for r in response["products"]:
-                        # Find beer
-                        beer = Beer.objects.get(vmp_id=int(r["code"]))
-                        stocked_beer.append(beer)
-
-                        quantity = int(
-                            r["availability"]["storeAvailability"]["mainText"].split()[
-                                1
-                            ]
+                # Update all beers in stock
+                for page in range(0, int(total_pages)):
+                    try:
+                        response, total_pages = call_api(
+                            url, store.store_id, page, product
                         )
 
-                        try:
-                            stock = Stock.objects.get(store=store, beer=beer)
-                            stock.quantity = quantity
-                            stock.save()
+                        for r in response["products"]:
+                            # Find beer
+                            beer = Beer.objects.get(vmp_id=int(r["code"]))
+                            stocked_beer.append(beer)
 
-                            updated += 1
-
-                        except Stock.DoesNotExist:
-                            stock = Stock.objects.create(
-                                store=store,
-                                beer=beer,
-                                quantity=quantity,
+                            quantity = int(
+                                r["availability"]["storeAvailability"][
+                                    "mainText"
+                                ].split()[1]
                             )
 
-                            created += 1
+                            try:
+                                stock = Stock.objects.get(store=store, beer=beer)
+                                stock.quantity = quantity
+                                stock.save()
 
-                except:
-                    continue
+                                updated += 1
+
+                            except Stock.DoesNotExist:
+                                stock = Stock.objects.create(
+                                    store=store,
+                                    beer=beer,
+                                    quantity=quantity,
+                                )
+
+                                created += 1
+
+                    except:
+                        continue
 
             # Remove all beers no longer in stock in the store
-            stocks = Stock.objects.filter(store=store).exclude(beer__in=stocked_beer)
-            deleted += stocks.count()
-            stocks.delete()
+            if len(stocked_beer) != 0:
+                stocks = Stock.objects.filter(store=store).exclude(
+                    beer__in=stocked_beer
+                )
+                deleted += stocks.count()
+                stocks.delete()
 
             store.store_stock_updated = timezone.now()
             store.save()
